promote checkstyle to a default report
checkstyle will run with the AddThis style on all builds, indivdual
projects will no longer need copy and paste.

Note: The least annoying thing to do seems to be to link to the
checkstyle file in master.  It changes rarely and linking to a tag is
a pain when developing new versions.  This does mean that changes in
master can affect the outcome of tagged downstream builds in a minor
way.